+++ xknx/devices/climate_mode.py
@@ -8,17 +8,17 @@ Controller modes use DPT 20.105.
 from __future__ import annotations
 
 from collections.abc import Iterator
-from itertools import chain
 from typing import TYPE_CHECKING, Any
 
-from xknx.dpt.dpt_20 import HVACControllerMode, HVACOperationMode
+from xknx.dpt.dpt_20 import HVACControllerMode, HVACOperationMode, HVACStatus
 from xknx.exceptions import DeviceIllegalValue
-from xknx.remote_value import GroupAddressesType, RemoteValue
+from xknx.remote_value import GroupAddressesType
 from xknx.remote_value.remote_value_climate_mode import (
     RemoteValueBinaryHeatCool,
     RemoteValueBinaryOperationMode,
     RemoteValueClimateModeBase,
     RemoteValueControllerMode,
+    RemoteValueControllerStatus,
     RemoteValueOperationMode,
 )
 
@@ -63,8 +63,7 @@ class ClimateMode(Device):
             sync_state=sync_state,
             device_name=name,
             feature_name="Operation mode",
-            climate_mode_type=RemoteValueOperationMode.ClimateModeType.HVAC_MODE,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_operation_mode,
         )
         self.remote_value_controller_mode = RemoteValueControllerMode(
             xknx,
@@ -73,17 +72,16 @@ class ClimateMode(Device):
             sync_state=sync_state,
             device_name=name,
             feature_name="Controller mode",
-            after_update_cb=None,
+            after_update_cb=self._set_internal_controller_mode,
         )
-        self.remote_value_controller_status = RemoteValueOperationMode(
+        self.remote_value_controller_status = RemoteValueControllerStatus(
             xknx,
             group_address=group_address_controller_status,
             group_address_state=group_address_controller_status_state,
             sync_state=sync_state,
             device_name=name,
             feature_name="Controller status",
-            climate_mode_type=RemoteValueOperationMode.ClimateModeType.CONTROLLER_STATUS,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_modes_from_status,
         )
 
         self.remote_value_operation_mode_comfort = RemoteValueBinaryOperationMode(
@@ -94,7 +92,7 @@ class ClimateMode(Device):
             device_name=name,
             feature_name="Operation mode Comfort",
             operation_mode=HVACOperationMode.COMFORT,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_operation_mode,
         )
         self.remote_value_operation_mode_standby = RemoteValueBinaryOperationMode(
             xknx,
@@ -104,7 +102,7 @@ class ClimateMode(Device):
             device_name=name,
             feature_name="Operation mode Standby",
             operation_mode=HVACOperationMode.STANDBY,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_operation_mode,
         )
         self.remote_value_operation_mode_night = RemoteValueBinaryOperationMode(
             xknx,
@@ -114,7 +112,7 @@ class ClimateMode(Device):
             device_name=name,
             feature_name="Operation mode Night",
             operation_mode=HVACOperationMode.NIGHT,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_operation_mode,
         )
         self.remote_value_operation_mode_protection = RemoteValueBinaryOperationMode(
             xknx,
@@ -124,7 +122,7 @@ class ClimateMode(Device):
             device_name=name,
             feature_name="Operation mode Protection",
             operation_mode=HVACOperationMode.FROST_PROTECTION,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_operation_mode,
         )
         self.remote_value_heat_cool = RemoteValueBinaryHeatCool(
             xknx,
@@ -134,7 +132,7 @@ class ClimateMode(Device):
             device_name=name,
             feature_name="Heat/Cool",
             controller_mode=HVACControllerMode.HEAT,
-            after_update_cb=None,
+            after_update_cb=self._set_internal_controller_mode,
         )
 
         self.operation_mode = HVACOperationMode.STANDBY
@@ -160,66 +158,27 @@ class ClimateMode(Device):
                 elif isinstance(ct_mode, HVACControllerMode):
                     self._controller_modes.append(ct_mode)
 
-        self.supports_operation_mode = any(
-            operation_mode.initialized
-            for operation_mode in self._iter_operation_remote_values()
-        )
-        self.supports_controller_mode = any(
-            controller_mode.initialized
-            for controller_mode in self._iter_controller_remote_values()
-        )
-        self._use_binary_operation_modes = any(
-            operation_mode.initialized
-            for operation_mode in self._iter_binary_operation_modes()
-        )
+        self.supports_operation_mode = bool(self._operation_modes)
+        self.supports_controller_mode = bool(self._controller_modes)
 
     def _iter_remote_values(
         self,
-    ) -> Iterator[RemoteValue[Any]]:
+    ) -> Iterator[RemoteValueClimateModeBase[Any]]:
         """Iterate climate mode RemoteValue classes."""
-        return chain(
-            self._iter_operation_remote_values(),
-            self._iter_controller_remote_values(),
-        )
-
-    def _iter_operation_remote_values(
-        self,
-    ) -> Iterator[RemoteValueClimateModeBase[HVACOperationMode]]:
-        return chain(
-            self._iter_binary_operation_modes(),
-            self._iter_byte_operation_modes(),
-        )
-
-    def _iter_binary_operation_modes(
-        self,
-    ) -> Iterator[RemoteValueClimateModeBase[HVACOperationMode]]:
-        """Iterate DPT 1 binary operation modes."""
-        yield from (
-            self.remote_value_operation_mode_comfort,
-            self.remote_value_operation_mode_night,
-            self.remote_value_operation_mode_protection,
-            self.remote_value_operation_mode_standby,
-        )
-
-    def _iter_byte_operation_modes(
-        self,
-    ) -> Iterator[RemoteValueClimateModeBase[HVACOperationMode]]:
-        """Iterate normal DPT 20.102 operation mode remote values."""
-        yield from (
-            self.remote_value_operation_mode,
-            self.remote_value_controller_status,
-        )
-
-    def _iter_controller_remote_values(
-        self,
-    ) -> Iterator[RemoteValueClimateModeBase[HVACControllerMode]]:
-        """Iterate DPT 20.105 controller remote values."""
+        yield self.remote_value_operation_mode
         yield self.remote_value_controller_mode
+        yield self.remote_value_controller_status
         yield self.remote_value_heat_cool
+        yield self.remote_value_operation_mode_comfort
+        yield self.remote_value_operation_mode_night
+        yield self.remote_value_operation_mode_protection
+        yield self.remote_value_operation_mode_standby
 
-    def _set_internal_operation_mode(self, operation_mode: HVACOperationMode) -> None:
+    def _set_internal_operation_mode(
+        self, operation_mode: HVACOperationMode | None
+    ) -> None:
         """Set internal value of operation mode. Call hooks if operation mode was changed."""
-        if operation_mode != self.operation_mode:
+        if operation_mode is not None and operation_mode != self.operation_mode:
             self.operation_mode = operation_mode
             self.after_update()
 
@@ -229,6 +188,18 @@ class ClimateMode(Device):
         """Set internal value of controller mode. Call hooks if controller mode was changed."""
         if controller_mode != self.controller_mode:
             self.controller_mode = controller_mode
+            self.after_update()
+
+    def _set_internal_modes_from_status(self, status: HVACStatus) -> None:
+        """Set internal values from HVACStatus."""
+        updated = False
+        if status.mode != self.operation_mode:
+            self.operation_mode = status.mode
+            updated = True
+        if status.heat_cool != self.controller_mode:
+            self.controller_mode = status.heat_cool
+            updated = True
+        if updated:
             self.after_update()
 
     async def set_operation_mode(self, operation_mode: HVACOperationMode) -> None:
@@ -241,13 +212,9 @@ class ClimateMode(Device):
                 "operation (preset) mode not supported", str(operation_mode)
             )
 
-        rv_operation: RemoteValueClimateModeBase[HVACOperationMode]
-        for rv_operation in self._iter_operation_remote_values():
-            if (
-                rv_operation.writable
-                and operation_mode in rv_operation.supported_operation_modes()
-            ):
-                rv_operation.set(operation_mode)
+        for rv in self._iter_remote_values():
+            if rv.writable:
+                rv.set_operation_mode(operation_mode)
 
         self._set_internal_operation_mode(operation_mode)
 
@@ -261,13 +228,9 @@ class ClimateMode(Device):
                 "controller (HVAC) mode not supported", str(controller_mode)
             )
 
-        rv_controller: RemoteValueClimateModeBase[HVACControllerMode]
-        for rv_controller in self._iter_controller_remote_values():
-            if (
-                rv_controller.writable
-                and controller_mode in rv_controller.supported_operation_modes()
-            ):
-                rv_controller.set(controller_mode)
+        for rv in self._iter_remote_values():
+            if rv.writable:
+                rv.set_controller_mode(controller_mode)
 
         self._set_internal_controller_mode(controller_mode)
 
@@ -288,37 +251,25 @@ class ClimateMode(Device):
     def gather_operation_modes(self) -> list[HVACOperationMode]:
         """Gather operation modes from RemoteValues."""
         operation_modes: list[HVACOperationMode] = []
-        for rv_operation in self._iter_operation_remote_values():
-            if rv_operation.writable:
-                operation_modes.extend(rv_operation.supported_operation_modes())
+        for rv in self._iter_remote_values():
+            if rv.writable:
+                operation_modes.extend(rv.supported_operation_modes())
         # remove duplicates
         return list(set(operation_modes))
 
     def gather_controller_modes(self) -> list[HVACControllerMode]:
         """Gather controller modes from RemoteValues."""
         controller_modes: list[HVACControllerMode] = []
-        for rv_controller in self._iter_controller_remote_values():
-            if rv_controller.writable:
-                controller_modes.extend(rv_controller.supported_operation_modes())
+        for rv in self._iter_remote_values():
+            if rv.writable:
+                controller_modes.extend(rv.supported_controller_modes())
         # remove duplicates
         return list(set(controller_modes))
 
     def process_group_write(self, telegram: Telegram) -> None:
         """Process incoming and outgoing GROUP WRITE telegram."""
-        if self.supports_operation_mode:
-            for rv_mode in self._iter_operation_remote_values():
-                if rv_mode.process(telegram):
-                    #  ignore inactive RemoteValueBinaryOperationMode
-                    if rv_mode.value:
-                        self._set_internal_operation_mode(rv_mode.value)
-                        return
-
-        if self.supports_controller_mode:
-            for rv_controller in self._iter_controller_remote_values():
-                if rv_controller.process(telegram):
-                    if rv_controller.value is not None:
-                        self._set_internal_controller_mode(rv_controller.value)
-                    return
+        for rv in self._iter_remote_values():
+            rv.process(telegram)
 
     def __str__(self) -> str:
         """Return object as readable string."""