+++ xknx/remote_value/remote_value_scaling.py
@@ -10,9 +10,8 @@ from typing import TYPE_CHECKING
 
 from xknx.dpt import DPTArray, DPTBinary
 from xknx.exceptions import CouldNotParseTelegram
-from xknx.typing import CallbackType
 
-from .remote_value import GroupAddressesType, RemoteValue
+from .remote_value import GroupAddressesType, RemoteValue, RVCallbackType
 
 if TYPE_CHECKING:
     from xknx.xknx import XKNX
@@ -29,7 +28,7 @@ class RemoteValueScaling(RemoteValue[int]):
         sync_state: bool | int | float | str = True,
         device_name: str | None = None,
         feature_name: str = "Value",
-        after_update_cb: CallbackType | None = None,
+        after_update_cb: RVCallbackType[int] | None = None,
         range_from: int = 0,
         range_to: int = 100,
     ):